Fixed a range of issues highlighted by IntelliJ Idea. Added LiveList - a list compatible object which supports change history tracking, incremental computation, and serialisation.

+++ java/BritefuryJ/DocModel/DMObjectOutputStream.java
@@ -60,7 +60,7 @@ public class DMObjectOutputStream
 			stream.writeBoolean( true );
 			stream.writeObject( modName );
 			stream.writeObject( schema.getLocation() );
-			stream.writeObject( (Integer)schema.getVersion() );
+			stream.writeObject( schema.getVersion() );
 		}
 		else
 		{
